Move MonadQuery instances to orphan module
For the same reasons as MonadTransaction, these instances are made
orphans so that they are opt-in. This is especially relevant because the
ReaderT instance is incompatible with `SqlPersistT`.
